修正
- 例の文章の修正
- エントリーポイントの分離 refs #6
